libdwfl: add interface for getting Dwfl_Module and Dwarf_Frame for Dwfl_Frame
The next change will need to have the Dwarf_Frame readily available, so
let's add dwfl_frame_dwarf_frame which looks up the Dwarf_Frame and
caches it. The CFI frame can also be useful to clients of libdwfl, so
let's make it a public interface. We can reorganize
__libdwfl_frame_unwind and handle_cfi to use it, too.

Similarly, the Dwfl_Module is frequently needed in conjuction with the
frame, so add dwfl_frame_module.